refactor(favorites): extract FavoritesFuelTab from favorites_screen (#388)
Pulls the 110-line `_buildFavoritesTab` helper out of favorites_screen.dart
into a standalone ConsumerWidget so the screen's build method becomes a
thin Scaffold + TabBar shell and so each render branch (empty / EV-only /
loaded / error) can be exercised in isolation.

favorites_screen.dart 199 -> 75 lines. Adds 3 widget tests covering the
empty branch, EV-only fallback, and error-with-retry branch — paths the
existing screen tests didn't reach because they only exercised the empty
and loaded states through the full Scaffold.

Closes part of #388.
